Fix for ingredient quantity

+++ app/src/main/java/com/sedsoftware/bakingapp/data/model/Ingredient.java
@@ -6,7 +6,7 @@ import com.google.gson.TypeAdapter;
 
 @AutoValue
 public abstract class Ingredient {
-  public abstract int quantity();
+  public abstract float quantity();
   public abstract String measure();
   public abstract String ingredient();
 
@@ -20,7 +20,7 @@ public abstract class Ingredient {
 
   @AutoValue.Builder
   public abstract static class Builder {
-    public abstract Builder quantity(int quantity);
+    public abstract Builder quantity(float quantity);
     public abstract Builder measure(String measure);
     public abstract Builder ingredient(String ingredient);
